+++ client/src/components/Nav/PcNav.tsx
@@ -1,15 +1,6 @@
 import tw from "tailwind-styled-components";
-import Image from "next/image";
-import {
-  // BsArrowLeftShort,
-  BsFillCaretDownFill,
-} from "react-icons/bs";
-import { useState } from "react";
-import Link from "next/link";
-import { useRouter } from "next/router";
-import { LoginMenus } from "./NavArr";
-import SearchBar from "./SearchBar";
-import logo from "../../image/logo.png";
+
+import NavContent from "./NavContent";
 
 export const NavBar = tw.nav`
     flex
@@ -43,76 +34,10 @@ hover:bg-white/30 rounded-md transitions duration-300 hover:text-white
 `;
 
 export default function Navbar() {
-  //   const [open, setOpen] = useState(true);
-  const [submenuOpen, setSubmenuOpen] = useState(false);
-  const router = useRouter();
-  const handleHomeClick = () => {
-    router.push("/");
-  };
   return (
     <NavBar>
-      {/* <BsArrowLeftShort
-        className={`bg-white text-dark-purple text-3xl rounded-full absolute right-3 top-9 border
-        border-dark-purple cursor-pointer ${!open && "rotate-180"}`}
-        role="button"
-        onClick={() => setOpen(!open)}
-      /> */}
-      <Nav className={`w-72 transitions duration-300 `}>
-        {/* ${open ? "left-0" : "left-[-72rem]"} */}
-        <Logo>
-          <Image
-            src={logo}
-            role="button"
-            alt="Home"
-            width={45}
-            onClick={handleHomeClick}
-          />
-        </Logo>
-        <SearchBar />
-        <NavList>
-          {LoginMenus.map(menu => (
-            <>
-              <Link href={menu.href}>
-                <List
-                  key={menu.id}
-                  className={`${
-                    router.pathname === menu.href && "text-yellow-200"
-                  }`}
-                >
-                  <span className="text-2xl block float-left">{menu.icon}</span>
-                  <span className="text-base font-medium flex-1 duration-200">
-                    {menu.title}
-                  </span>
-                  {menu.submenu && (
-                    <BsFillCaretDownFill
-                      size="25px"
-                      role="button"
-                      className={`transtions duration-300 ${
-                        submenuOpen && "rotate-180"
-                      }`}
-                      onClick={() => setSubmenuOpen(!submenuOpen)}
-                    />
-                  )}
-                </List>
-              </Link>
-              <SubNavList
-                className={`${
-                  menu.submenu && submenuOpen ? "translate-y-1" : ""
-                } duration-300 relative`}
-              >
-                {menu.submenu && submenuOpen && (
-                  <>
-                    {menu.submenuItems.map(submenuItem => (
-                      <SubList key={menu.id}>
-                        <Link href={submenuItem.href}>{submenuItem.title}</Link>
-                      </SubList>
-                    ))}
-                  </>
-                )}
-              </SubNavList>
-            </>
-          ))}
-        </NavList>
+      <Nav>
+        <NavContent />
       </Nav>
     </NavBar>
   );
